requirements.txt
fastapi==0.127.0
uvicorn[standard]==0.30.6

sqlalchemy==2.0.45
alembic==1.16.5
psycopg2-binary==2.9.11

pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
